moved TypeConverter and PendingListensDao to shared module with related DAO entities i.e:- AlbumEntity and SongEntity
