+++ app/src/main/java/org/listenbrainz/android/di/brainzplayer/BrainzPlayerDatabase.kt
@@ -5,15 +5,15 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
-import org.listenbrainz.android.model.AlbumEntity
+import org.listenbrainz.shared.model.AlbumEntity
 import org.listenbrainz.android.model.ArtistEntity
 import org.listenbrainz.android.model.PlaylistEntity
-import org.listenbrainz.android.model.SongEntity
+import org.listenbrainz.shared.model.SongEntity
 import org.listenbrainz.android.model.dao.AlbumDao
 import org.listenbrainz.android.model.dao.ArtistDao
 import org.listenbrainz.android.model.dao.PlaylistDao
 import org.listenbrainz.android.model.dao.SongDao
-import org.listenbrainz.android.util.TypeConverter
+import org.listenbrainz.shared.util.TypeConverter
 
 object Migrations {
     val MIGRATION_1_2: Migration = object : Migration(1, 2) {
