+++ app/src/main/java/org/listenbrainz/android/model/ArtistEntity.kt
@@ -3,6 +3,8 @@ package org.listenbrainz.android.model
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.serialization.Serializable
+import org.listenbrainz.shared.model.AlbumEntity
+import org.listenbrainz.shared.model.SongEntity
 
 @Serializable
 @Entity(tableName = "ARTISTS")
