+++ app/src/main/java/org/listenbrainz/android/model/dao/PlaylistDao.kt
@@ -3,7 +3,7 @@ package org.listenbrainz.android.model.dao
 import androidx.room.*
 import kotlinx.coroutines.flow.Flow
 import org.listenbrainz.android.model.PlaylistEntity
-import org.listenbrainz.android.model.SongEntity
+import org.listenbrainz.shared.model.SongEntity
 
 @Dao
 interface PlaylistDao {
